refactor: organizing into directories
