+++ articles/IsolatedFunctions/libs/AzureServiceBusLib/Core/ErrorMessages.cs
@@ -1,6 +1,6 @@
-﻿namespace AzureServiceBusLib.NewCore;
+﻿namespace AzureServiceBusLib.Core;
 
-public static class ErrorMessages
+internal static class ErrorMessages
 {
     public const string MessagePublishError = "An error occurred while publishing the message to the service bus.";
     public const string TooManyMessagesInBatch = "Too many messages in the batch";
