--- articles/IsolatedFunctions/libs/AzureServiceBusLib/NewCore/IServiceBusMessagePublisher.cs
@@ -1,6 +1,0 @@
-﻿namespace AzureServiceBusLib.NewCore;
-
-public interface IServiceBusMessagePublisher
-{
-    public string Name { get; }
-}
