+++ articles/IsolatedFunctions/tests/OrderPublisher.Console.Tests/ServiceBusPublisherTests.cs
@@ -1,7 +1,9 @@
 ﻿using System.Text.Json;
 using System.Text.Json.Serialization;
 using AutoBogus;
-using AzureServiceBusLib.NewCore;
+using AzureServiceBusLib.Core;
+using AzureServiceBusLib.Models;
+using AzureServiceBusLib.Services;
 using Microsoft.Extensions.DependencyInjection;
 using OrderPublisher.Console.Models;
