+++ articles/IsolatedFunctions/tools/OrderPublisher.Console/Bootstrapper.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text.Json;
 using System.Text.Json.Serialization;
-using AzureServiceBusLib.NewCore;
+using AzureServiceBusLib.Core;
+using AzureServiceBusLib.Services;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
